+++ libs/rxjs/number/src/lib/from-number.spec.ts
@@ -1,48 +1,29 @@
-import { catchError, map, reduce, tap } from 'rxjs/operators';
+import { map, reduce, take, tap } from 'rxjs/operators';
 import { fromNumber } from '@rxjs-ninja/rxjs-number';
 import { observe } from 'rxjs-marbles/jest';
 import { of } from 'rxjs';
 
 describe('fromNumber', () => {
   it(
-    'should create observable from number value',
+    'should create a sequence of numbers if no input passed, reduce to array, adding `1` to each value',
     observe(() =>
-      fromNumber(5).pipe(
-        map((val) => val * 2),
-        tap((value) => expect(value).toBe(10)),
-      ),
-    ),
-  );
-
-  it(
-    'should create an Error from a failed promise',
-    observe(() =>
-      fromNumber(Promise.reject('RxJS Ninja')).pipe(
-        map((val) => val * 2),
-        catchError((error) => {
-          expect(error).toBe('RxJS Ninja');
-          return of(true);
+      fromNumber().pipe(
+        take(100),
+        reduce((a, b) => [...a, b + 1], [] as number[]),
+        tap((value) => {
+          expect(value.length).toBe(100);
+          expect(value[0]).toBe(1);
         }),
       ),
     ),
   );
 
   it(
-    'should create observable from promise number value',
+    'should create observable from number value',
     observe(() =>
-      fromNumber(Promise.resolve(5)).pipe(
+      fromNumber(5).pipe(
         map((val) => val * 2),
         tap((value) => expect(value).toBe(10)),
-      ),
-    ),
-  );
-
-  it(
-    'should create observable from promise number list',
-    observe(() =>
-      fromNumber(Promise.resolve([5, 10, 20])).pipe(
-        reduce((a, b) => a + b, 0),
-        tap((value) => expect(value).toBe(35)),
       ),
     ),
   );
@@ -63,16 +44,6 @@ describe('fromNumber', () => {
       fromNumber(of([5, 10, 20])).pipe(
         reduce((a, b) => a + b, 0),
         tap((value) => expect(value).toBe(35)),
-      ),
-    ),
-  );
-
-  it(
-    'should create observable from argument list of number values',
-    observe(() =>
-      fromNumber(1, 2, 3, 4).pipe(
-        reduce((acc, val) => acc + val),
-        tap((value) => expect(value).toBe(10)),
       ),
     ),
   );