+++ libs/rxjs/number/src/lib/is-integer.ts
@@ -16,7 +16,7 @@ import { map } from 'rxjs/operators';
  * Return booleans for values that are integer values
  * ```ts
  * const input = [-10, -2.3, 0, 1, 2, 3.14, 4.2, 10, 11, 42];
- * from(input).pipe(filterIsInteger()).subscribe()
+ * from(input).pipe(isInteger()).subscribe()
  * ```
  * Output: `true, false, true, true, true, false, false, true, true, true`
  *