+++ libs/rxjs/number/src/lib/is-not-nan.ts
@@ -6,13 +6,15 @@ import { OperatorFunction } from 'rxjs';
 import { map } from 'rxjs/operators';
 
 /**
- * Returns an Observable that emits a boolean value when a source number is valid and not `NaN`, checked using Number.isNaN
+ * Returns an Observable that emits a boolean value when a source number is valid and not `NaN`, checked using
+ * Number.isNaN
  *
  * @category Query
  *
- * @remarks This is mostly for convenience of getting truthy values that are valid for numbers without flipping the value os `isNaN`
+ * @remarks This is mostly for convenience of getting truthy values that are valid for numbers without flipping the
+ *   value os `isNaN`
  *
- * @see The [[filterIsNotNaN]] operator returns the number value
+ * @see The [[filterNaN]] operator returns numbers excluding `NaN` values
  *
  * @example
  * Return a boolean if a number is not `NaN` value