+++ libs/rxjs/number/src/lib/parse-hex.spec.ts
@@ -7,19 +7,8 @@ describe('parseHex', () => {
     marbles((m) => {
       const input = m.hot('-a-b-c-d-e-|', { a: 'ff', b: '00', c: '1b', d: 'c89bb', e: 'Ninja' });
       const subs = '^----------!';
-      const expected = m.cold('-w-v-x-y---|', { w: 255, v: 0, x: 27, y: 821691 });
-      m.expect(input.pipe(parseHex())).toBeObservable(expected);
-      m.expect(input).toHaveSubscriptions(subs);
-    }),
-  );
-
-  it(
-    'should return parsed number values and return NaN values',
-    marbles((m) => {
-      const input = m.hot('-a-b-c-d-e-|', { a: 'ff', b: '00', c: '1b', d: 'c89bb', e: 'Ninja' });
-      const subs = '^----------!';
       const expected = m.cold('-w-v-x-y-z-|', { w: 255, v: 0, x: 27, y: 821691, z: NaN });
-      m.expect(input.pipe(parseHex(true))).toBeObservable(expected);
+      m.expect(input.pipe(parseHex())).toBeObservable(expected);
       m.expect(input).toHaveSubscriptions(subs);
     }),
   );