+++ libs/rxjs/number/src/lib/pow.ts
@@ -2,8 +2,9 @@
  * @packageDocumentation
  * @module Number
  */
-import { isObservable, MonoTypeOperatorFunction, Observable, ObservableInput, of } from 'rxjs';
+import { MonoTypeOperatorFunction, Subscribable } from 'rxjs';
 import { map, withLatestFrom } from 'rxjs/operators';
+import { createOrReturnObservable } from '../utils/internal';
 
 /**
  * Returns an Observable that emits a number from a source number that is raised by the passed power using the
@@ -16,14 +17,14 @@ import { map, withLatestFrom } from 'rxjs/operators';
  * @example Return values raised to the power `2`
  * ```ts
  * const input = [2, 4, 10, 16, 256];
- * from(input).pipe(raiseBy(2)).subscribe();
+ * from(input).pipe(pow(2)).subscribe();
  * ```
  * Output: `4, 25, 100, 256, 655356`
  *
  * @returns Observable that emits a number that is the raised source value by the power
  */
-export function pow(power: number | ObservableInput<number>): MonoTypeOperatorFunction<number> {
-  const power$ = (isObservable(power) ? power : of(power)) as Observable<number>;
+export function pow(power: Subscribable<number> | number): MonoTypeOperatorFunction<number> {
+  const power$ = createOrReturnObservable(power);
   return (source) =>
     source.pipe(
       withLatestFrom(power$),