+++ libs/rxjs/number/src/lib/to-hex.ts
@@ -7,7 +7,9 @@ import { map } from 'rxjs/operators';
 
 /**
  * Returns an Observable that emits a string that is the hex value of a source number
- **
+ *
+ * @category Formatting
+ *
  * @example Return a hex value of source number
  * ```ts
  * const input = [2, 16, 32, 75, 255];
@@ -16,7 +18,6 @@ import { map } from 'rxjs/operators';
  * Output: `'2', '10', '20', '4b', 'ff'`
  *
  * @returns Observable that emits the hex value of a source number
- * @category Formatting
  */
 export function toHex(): OperatorFunction<number, string> {
   return (source: Observable<number>) => source.pipe(map((number) => number.toString(16)));