+++ libs/rxjs/number/src/lib/to-locale-string.ts
@@ -2,12 +2,15 @@
  * @packageDocumentation
  * @module Number
  */
-import { isObservable, Observable, ObservableInput, of, OperatorFunction } from 'rxjs';
+import { Observable, OperatorFunction, Subscribable } from 'rxjs';
 import { map, withLatestFrom } from 'rxjs/operators';
+import { createOrReturnObservable } from '../utils/internal';
 
 /**
  * Returns an Observable that emits a formatted string value from a source number using Number.toLocaleString with
- * optional formatting options provided by Intl.NumberFormat.
+ * optional formatting options provided by Intl.NumberFormat
+ *
+ * @category Formatting
  *
  * @param locales The locale or locales that the number is being formatted for
  * @param format Formatting of the string based on
@@ -32,20 +35,18 @@ import { map, withLatestFrom } from 'rxjs/operators';
  * Output: `'€1,000,000.00'`
  *
  * @returns Observable that emits a formatted string from a source number
- * @category Formatting
  */
 export function toLocaleString(
-  locales?: ObservableInput<string | string[]> | string | string[],
-  format?: ObservableInput<Intl.NumberFormatOptions> | Intl.NumberFormatOptions,
+  locales?: Subscribable<string | string[]> | string | string[],
+  format?: Subscribable<Intl.NumberFormatOptions> | Intl.NumberFormatOptions,
 ): OperatorFunction<number, string> {
-  const locales$ = (isObservable(locales) ? locales : of(locales)) as Observable<string | string[]>;
-  const format$ = (isObservable(format) ? format : of(format)) as Observable<Intl.NumberFormatOptions>;
-
+  const locales$ = createOrReturnObservable(locales);
+  const format$ = createOrReturnObservable(format);
   return (source: Observable<number>) =>
     source.pipe(
       withLatestFrom(locales$, format$),
-      map<[number, string | string[], Intl.NumberFormatOptions], string>(([number, localeValue, formatValue]) =>
-        number.toLocaleString(localeValue, formatValue),
+      map<[number, string | string[] | undefined, Intl.NumberFormatOptions | undefined], string>(
+        ([number, localeValue, formatValue]) => number.toLocaleString(localeValue, formatValue),
       ),
     );
 }