+++ libs/rxjs/number/src/lib/to-precision.ts
@@ -2,11 +2,14 @@
  * @packageDocumentation
  * @module Number
  */
-import { isObservable, Observable, ObservableInput, of, OperatorFunction } from 'rxjs';
+import { OperatorFunction, Subscribable } from 'rxjs';
 import { map, withLatestFrom } from 'rxjs/operators';
+import { createOrReturnObservable } from '../utils/internal';
 
 /**
  * Returns an Observable that emits a formatted string value from a source number using Number.toPrecision.
+ *
+ * @category Formatting
  *
  * @param precision The number of decimal places to format the precision to.
  *
@@ -18,13 +21,12 @@ import { map, withLatestFrom } from 'rxjs/operators';
  * Output: `'123.5', '0.004000', '1.230e+5'`
  *
  * @returns Observable that emits a formatted string from a source number
- * @category Formatting
  */
-export function toPrecision(precision: number | ObservableInput<number>): OperatorFunction<number, string> {
-  const precision$ = (isObservable(precision) ? precision : of(precision)) as Observable<number>;
+export function toPrecision(precision: Subscribable<number> | number): OperatorFunction<number, string> {
+  const precision$ = createOrReturnObservable(precision);
   return (source) =>
     source.pipe(
       withLatestFrom(precision$),
-      map<[number, number], string>(([value, inputValue]) => value.toPrecision(inputValue)),
+      map(([value, precisionValue]) => value.toPrecision(precisionValue)),
     );
 }